Move testutil package out of internal
