+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+Move `testutil` package out of `internal`
+
 ## [0.9.10] - 2020-08-20
 ### Added
 - The `Resource` field was added to Entry ([PR95](https://github.com/observIQ/carbon/pull/95))
