+++ agent/agent_test.go
@@ -6,7 +6,7 @@ import (
 	"runtime"
 	"testing"
 
-	"github.com/observiq/carbon/internal/testutil"
+	"github.com/observiq/carbon/testutil"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 )
