+++ operator/builtin/output/stdout_test.go
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	"github.com/observiq/carbon/entry"
-	"github.com/observiq/carbon/internal/testutil"
 	"github.com/observiq/carbon/operator/helper"
+	"github.com/observiq/carbon/testutil"
 	"github.com/stretchr/testify/require"
 )
