+++ operator/builtin/parser/json_test.go
@@ -7,9 +7,9 @@ import (
 
 	jsoniter "github.com/json-iterator/go"
 	"github.com/observiq/carbon/entry"
-	"github.com/observiq/carbon/internal/testutil"
 	"github.com/observiq/carbon/operator"
 	"github.com/observiq/carbon/operator/helper"
+	"github.com/observiq/carbon/testutil"
 
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
