+++ operator/builtin/parser/regex_test.go
@@ -6,9 +6,9 @@ import (
 	"testing"
 
 	"github.com/observiq/carbon/entry"
-	"github.com/observiq/carbon/internal/testutil"
 	"github.com/observiq/carbon/operator"
 	"github.com/observiq/carbon/operator/helper"
+	"github.com/observiq/carbon/testutil"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 )
