+++ operator/builtin/transformer/noop_test.go
@@ -3,7 +3,7 @@ package transformer
 import (
 	"testing"
 
-	"github.com/observiq/carbon/internal/testutil"
+	"github.com/observiq/carbon/testutil"
 	"github.com/stretchr/testify/require"
 )
