+++ operator/helper/input_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/observiq/carbon/entry"
-	"github.com/observiq/carbon/internal/testutil"
+	"github.com/observiq/carbon/testutil"
 	"github.com/stretchr/testify/require"
 )
