+++ operator/helper/operator_test.go
@@ -3,8 +3,8 @@ package helper
 import (
 	"testing"
 
-	"github.com/observiq/carbon/internal/testutil"
 	"github.com/observiq/carbon/operator"
+	"github.com/observiq/carbon/testutil"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 )
