+++ pipeline/pipeline_test.go
@@ -3,8 +3,8 @@ package pipeline
 import (
 	"testing"
 
-	"github.com/observiq/carbon/internal/testutil"
 	"github.com/observiq/carbon/operator"
+	"github.com/observiq/carbon/testutil"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"gonum.org/v1/gonum/graph"
